+++ include/janus/benchmark.hpp
@@ -4,7 +4,6 @@
 
 namespace janus {
 
-void self_test();
 void benchmark(std::size_t message_count, bool header = true);
 
 }  // namespace janus