+++ include/janus/update.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace janus {
+
+void update();
+
+}  // namespace janus